types fixed

+++ hash.h
@@ -8,7 +8,7 @@ struct file_hash {
 	int count;
 };
 
-long long integer_mod(int, long long);
+long long integer_mod(long long, long long);
 struct file_hash* generate_hashes(const char*, int);
 char* substring(int start, int, char*);
 char** generate_k_grams(char*, int);
